+++ src/move_ordering.h
@@ -24,14 +24,14 @@ Score score_q_bn(Thread_State& thread_state, Move move, Move tt_move,
                       InformativeMove last_moves[]);
 
 template<bool qsearch>
-Score score_capture(Thread_State& thread_state, ScoredMove& scored_move, Move tt_move, int& good_capture_count);
+Score score_capture(Thread_State& thread_state, ScoredMove& scored_move, Move tt_move, size_t& good_capture_count);
 
 void get_q_bn_scores(Thread_State& thread_state, FixedVector<ScoredMove, MAX_MOVES>& current_scored_moves,
                      Move tt_move, InformativeMove last_moves[], int start_index);
 
 template<bool qsearch>
 void get_capture_scores(Thread_State& thread_state, FixedVector<ScoredMove, MAX_MOVES>& current_scored_moves,
-                        Move tt_move, int& good_capture_count);
+                        Move tt_move, size_t& good_capture_count);
 
 enum class Filter : uint16_t {
     None,
@@ -66,9 +66,9 @@ public:
     InformativeMove last_moves[LAST_MOVE_COUNTS]{};
 
     int stage = Stage::TT_probe;
-    int move_index = 0;
-    int good_capture_count = 0;
-    int good_capture_found = 0;
+    size_t move_index = 0;
+    size_t good_capture_count = 0;
+    size_t good_capture_found = 0;
 
     Ply search_ply = 0;
